Treat Cygwin as a separate OS
Cygwin is considered a distinct target with a distinct ABI, environment
conditions, and data types.  Though the goal of the project is
native Windows integration with UNIX-likes, that is not compatible with
the idea that the platform can be ignored as Win-like enough to have the
existing os(Windows) condition apply.

# Conflicts:
#	cmake/modules/SwiftSource.cmake
#	lib/Basic/LangOptions.cpp
#	stdlib/private/StdlibUnittest/RaceTest.swift
#	stdlib/private/StdlibUnittest/StdlibCoreExtras.swift
#	stdlib/private/SwiftPrivateLibcExtras/Subprocess.swift
#	stdlib/private/SwiftPrivateLibcExtras/SwiftPrivateLibcExtras.swift
#	stdlib/private/SwiftPrivatePthreadExtras/PthreadBarriers.swift
#	stdlib/private/SwiftPrivatePthreadExtras/SwiftPrivatePthreadExtras.swift
#	stdlib/public/Platform/Platform.swift
#	stdlib/public/Platform/tgmath.swift.gyb
#	test/Parse/ConditionalCompilation/x64CygwinTarget.swift